+++ actix-server/src/test_server.rs
@@ -103,10 +103,10 @@ impl TestServer {
         let port = addr.port();
 
         TestServerRuntime {
-            system,
             addr,
             host,
             port,
+            system,
         }
     }