+++ actix-tls/CHANGES.md
@@ -13,7 +13,7 @@
 [#295]: https://github.com/actix/actix-net/pull/295
 [#296]: https://github.com/actix/actix-net/pull/296
 [#297]: https://github.com/actix/actix-net/pull/297
-[#299]: https://github.com/actix/actix-net/pull/298
+[#299]: https://github.com/actix/actix-net/pull/299
 
 
 ## 3.0.0-beta.4 - 2021-02-24